+++ CHANGES.md
@@ -3,6 +3,7 @@ unreleased
 
   + merlin library
     - Implement new refactor-extract-region command for extracting region to a fresh let binding (#1546)
+    - Fix `merlin_reader` for OpenBSD (#1956)
 
 merlin 5.5
 ==========