+++ CHANGES.md
@@ -37,6 +37,7 @@ Tue Jun 24 16:10:42 CEST 2025
       of the involved data-structures. (#1889)
   + test suite
     - Add a test case illustrating wrong open order proposed in issue #1900. (#1901)
+    - Add more short-paths tests cases (#1904)
 
 merlin 5.4.1
 ============